refactor(builder): move InstrumentedFinishProvider to metrics.rs
Amp-Thread-ID: https://ampcode.com/threads/T-019cd693-2c26-70a0-ad90-da4435a2d6c7
