+++ crates/payload/builder/src/metrics.rs
@@ -1,5 +1,18 @@
+use alloy_primitives::{Address, B256, BlockNumber, Bytes, StorageKey, StorageValue};
 use metrics::Gauge;
+use reth_errors::ProviderResult;
 use reth_metrics::{Metrics, metrics::Histogram};
+use reth_primitives_traits::{Account, Bytecode};
+use reth_storage_api::{
+    AccountReader, BlockHashReader, BytecodeReader, HashedPostStateProvider, StateProofProvider,
+    StateProvider, StateRootProvider, StorageRootProvider,
+};
+use reth_trie_common::{
+    AccountProof, HashedPostState, HashedStorage, MultiProof, MultiProofTargets, StorageMultiProof,
+    StorageProof, TrieInput, updates::TrieUpdates,
+};
+use std::time::Instant;
+use tracing::debug_span;
 
 #[derive(Metrics, Clone)]
 #[metrics(scope = "tempo_payload_builder")]
@@ -79,4 +92,88 @@ pub(crate) struct TempoPayloadBuilderMetrics {
 pub(crate) fn inc_pool_tx_skipped(reason: &'static str) {
     metrics::counter!("tempo_payload_builder_pool_transactions_skipped_total", "reason" => reason)
         .increment(1);
+}
+
+/// Wraps a [`StateProvider`] reference to instrument `hashed_post_state` and
+/// `state_root_with_updates` with tracing spans and histogram metrics during `builder.finish()`.
+pub(crate) struct InstrumentedFinishProvider<'a> {
+    pub(crate) inner: &'a dyn StateProvider,
+    pub(crate) metrics: TempoPayloadBuilderMetrics,
+}
+
+impl<'a> AsRef<dyn StateProvider + 'a> for InstrumentedFinishProvider<'a> {
+    fn as_ref(&self) -> &(dyn StateProvider + 'a) {
+        self.inner
+    }
+}
+
+reth_storage_api::delegate_impls_to_as_ref!(
+    for InstrumentedFinishProvider<'_> =>
+    AccountReader {
+        fn basic_account(&self, address: &Address) -> ProviderResult<Option<Account>>;
+    }
+    BlockHashReader {
+        fn block_hash(&self, number: u64) -> ProviderResult<Option<B256>>;
+        fn canonical_hashes_range(&self, start: BlockNumber, end: BlockNumber) -> ProviderResult<Vec<B256>>;
+    }
+    StateProvider {
+        fn storage(&self, account: Address, storage_key: StorageKey) -> ProviderResult<Option<StorageValue>>;
+    }
+    BytecodeReader {
+        fn bytecode_by_hash(&self, code_hash: &B256) -> ProviderResult<Option<Bytecode>>;
+    }
+    StorageRootProvider {
+        fn storage_root(&self, address: Address, storage: HashedStorage) -> ProviderResult<B256>;
+        fn storage_proof(&self, address: Address, slot: B256, storage: HashedStorage) -> ProviderResult<StorageProof>;
+        fn storage_multiproof(&self, address: Address, slots: &[B256], storage: HashedStorage) -> ProviderResult<StorageMultiProof>;
+    }
+    StateProofProvider {
+        fn proof(&self, input: TrieInput, address: Address, slots: &[B256]) -> ProviderResult<AccountProof>;
+        fn multiproof(&self, input: TrieInput, targets: MultiProofTargets) -> ProviderResult<MultiProof>;
+        fn witness(&self, input: TrieInput, target: HashedPostState) -> ProviderResult<Vec<Bytes>>;
+    }
+);
+
+impl HashedPostStateProvider for InstrumentedFinishProvider<'_> {
+    fn hashed_post_state(&self, bundle_state: &reth_revm::db::BundleState) -> HashedPostState {
+        let start = Instant::now();
+        let _span = debug_span!(target: "payload_builder", "hashed_post_state").entered();
+        let result = self.inner.hashed_post_state(bundle_state);
+        drop(_span);
+        self.metrics
+            .hashed_post_state_duration_seconds
+            .record(start.elapsed());
+        result
+    }
+}
+
+impl StateRootProvider for InstrumentedFinishProvider<'_> {
+    fn state_root(&self, hashed_state: HashedPostState) -> ProviderResult<B256> {
+        self.inner.state_root(hashed_state)
+    }
+
+    fn state_root_from_nodes(&self, input: TrieInput) -> ProviderResult<B256> {
+        self.inner.state_root_from_nodes(input)
+    }
+
+    fn state_root_with_updates(
+        &self,
+        hashed_state: HashedPostState,
+    ) -> ProviderResult<(B256, TrieUpdates)> {
+        let start = Instant::now();
+        let _span = debug_span!(target: "payload_builder", "state_root_with_updates").entered();
+        let result = self.inner.state_root_with_updates(hashed_state);
+        drop(_span);
+        self.metrics
+            .state_root_with_updates_duration_seconds
+            .record(start.elapsed());
+        result
+    }
+
+    fn state_root_from_nodes_with_updates(
+        &self,
+        input: TrieInput,
+    ) -> ProviderResult<(B256, TrieUpdates)> {
+        self.inner.state_root_from_nodes_with_updates(input)
+    }
 }
